+++ Changelog.md
@@ -5,11 +5,13 @@ All notable changes to this project will be documented in this file since v2.0.0
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.2.4] - future release
+## [2.3.0] - 2021-07-23
 
 ### Added
 
 - `dockerize` - ask for confirmation if project root is outside the directory defined in the `PROJECTS_ROOT_DIR` environment variable.
+- `\App\Services\SslCErtificate` - new class responsible for generating SSL certificates via `mkcert`.
+- show full `dockerize` command after entering all parameters for reference and for the future re-use if needed.
 
 ### Changed